@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.5+6
+version: 1.0.6+7
 
 environment:
-  sdk: ">=2.17.5 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,22 +29,22 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  curved_navigation_bar: ^1.0.3
-  provider: ^6.0.3
-  shared_preferences: ^2.0.15
-  calendar_timeline: ^1.0.4
-  intl: ^0.18.1
+  curved_navigation_bar: ^1.0.6
+  provider: ^6.1.2
+  shared_preferences: ^2.5.1
+  calendar_timeline: ^1.1.3
+  intl: ^0.19.0
   validators: ^3.0.0
-  font_awesome_flutter: ^10.1.0
-  path: ^1.8.1
-  sqflite: ^2.0.3
-  uuid: ^3.0.6
-  flutter_typeahead: ^4.0.0
+  font_awesome_flutter: ^10.8.0
+  path: ^1.9.0
+  sqflite: ^2.4.1
+  uuid: ^4.5.1
+  flutter_typeahead: ^5.2.0
   hexcolor: ^3.0.1
   flip_card: ^0.7.0
   community_material_icon: ^5.9.55
-  jiffy: ^6.2.1
-  syncfusion_flutter_charts: ^22.2.9
+  jiffy: ^6.3.2
+  syncfusion_flutter_charts: ^28.2.3
   duration_picker: ^1.2.0
   # firebase_auth: ^3.6.3
   # firebase_core: ^1.23.0
@@ -53,13 +53,13 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.13.1
-  flutter_lints: ^2.0.1
+  flutter_launcher_icons: ^0.14.3
+  flutter_lints: ^5.0.0
 
 flutter_icons:
   android: true
